Remove kernel_interface.arrayobj
    - Merges core.kernel_interface.arrayobj into
      kernel_api_impl.spirv.arrayobj
    - Updated populate_array to use USMNdArray type.
    - Remove core.kernel_api
