--- numba_dpex/core/kernel_interface/__init__.py
@@ -1,6 +1,0 @@
-# SPDX-FileCopyrightText: 2020 - 2024 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""Defines the interface for kernel compilation using numba-dpex.
-"""
